+++ apps/plugins/mpegplayer/disk_buf.c
@@ -786,12 +786,12 @@ ssize_t disk_buf_prepare_streaming(off_t pos, size_t len)
     else if (pos > disk_buf.filesize)
         pos = disk_buf.filesize;
 
-    DEBUGF("prepare streaming:\n  pos:%ld len:%zu\n", pos, len);
+    DEBUGF("prepare streaming:\n  pos:%ld len:%lu\n", pos, (unsigned long)len);
 
     pos = disk_buf_lseek(pos, SEEK_SET);
     len = disk_buf_probe(pos, len, NULL);
 
-    DEBUGF("  probe done: pos:%ld len:%zu\n", pos, len);
+    DEBUGF("  probe done: pos:%ld len:%lu\n", pos, (unsigned long)len);
 
     len = disk_buf_send_msg(STREAM_RESET, pos);
